requirements.txt
alembic==1.4.2
appdirs==1.4.3
bleach==3.1.5
blinker==1.4
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.1.1
Flask-Admin==1.5.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==0.4.15
gunicorn==18.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==0.24
Jinja2==2.10.3
lockfile==0.12.2
Mako==1.1.2
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
pyparsing==2.4.6
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.17
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.3.1
